2008-12-06  Sebastien Pouliot  <sebastien@ximian.com>
	* TextResultWriter.cs: Big memory reduction in the LINQ query
	by casting the IComparable enum into an IComparable<T> Int32. Even
	bigger now that the query is ordered on two fields. Not sure if
	this has the same effect on MS LINQ or just on Mono.


svn path=/trunk/mono-tools/; revision=120952

+++ gendarme/console/TextResultWriter.cs
@@ -79,8 +79,9 @@ namespace Gendarme {
 
 		protected override void Write ()
 		{
+			// casting to Severity int saves a ton of memory since IComparable<T> can be used instead of IComparable
 			var query = from n in Runner.Defects
-				    orderby n.Severity, n.Rule.Name
+				    orderby (int) n.Severity, n.Rule.Name
 				    select n;
 			
 			WriteHeader ();
